moves SignUp component to new component, sets up firebase config with .env.local git ignored
