+++ src/firebase.js
@@ -1,0 +1,15 @@
+import firebase from 'firebase/app';
+import 'firebase/auth';
+
+const app = firebase.initialiseApp({
+  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
+  authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
+  projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
+  storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
+  appId: process.env.REACT_APP_FIREBASE_APP_ID
+  // databaseUrl not in config on firebase app -lewis
+});
+
+export const auth = app.auth();
+export default app;
